Update jasperserver-dsl.rb
At the company I work for, we have our reports organized in a hierarchical structure, for example.

- /accountability/bank1/income
- /accountability/bank2/income
- /courses/invoice
- /courses/grade_list

The code in this gem assumes every report is inside the ```reports``` folder, in our case that is not true. In order to use this gem we would have to move all our reports to the ```reports``` folder, for example.

- /reports/accountability/bank1/income
- /reports/accountability/bank2/income
- /reports/courses/invoice
- /reports/courses/grade_list

This is possible but impractical, we also would have to change every report call in all our codebase.

With this patch I propose to remove the hardcoded ```reports``` folder. Also, I assume someone else could benefit from this, which I think is a more general use case.